separated components, use memoization for post count calculation, simplified some code
